+++ Game/Models/Telecaster/PartyEntity.cs
@@ -7,12 +7,13 @@ namespace Navislamia.Game.Models.Telecaster;
 public class PartyEntity : Entity
 {
     public string Name { get; set; }
-    public int LeaderId { get; set; }
+    public long LeaderId { get; set; }
     public virtual CharacterEntity Leader { get; set; }
 
     public PartyItemShareMode ShareMode { get; set; }
     public PartyType PartyType { get; set; }
-    public int LeadPartyId { get; set; }
+    
+    public long LeadPartyId { get; set; }
     public virtual PartyEntity LeadParty { get; set; } // Attackteams/raids
 
     public virtual ICollection<CharacterEntity> PartyMembers { get; set; }
